Lets start this over again
This reorganizes GitView in preparation for the full rewrite.  This
is going to become the testbed for jsonapi support in Django REST
Framework, along with Python 3 support in GitPython.  Once the API
is established and tests are written for it, the frontend will be
built using whatever technology needs testing at the time.

In the past we experimented with AngularJS and determined that it
would not work well for the page count and complex templates that
we needed.  EmberJS was also evaluated, but not for a long enough
time to determine whether or not it would work with an application
of this size.

This will also use Django 1.7 along with Docker builds that will
allow us to have consistent environments that can be reproduced
easily.

A lot of files which are not going to be used, or that will be
completely rewritten, have been dropped for now.